refactor(agent): split CLI command domains
(cherry picked from commit b4983e25c167bf6f0cbe3bffd847d0f34c3f36c6)
