+++ framework/core/src/python/kungfu/cli/commands/_agent/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""Internal command domains for the stable ``kungfu agent`` facade."""
